Refactor SQL query for price per kg calculation
Updated SQL query to include type casting for price and weight.

+++ GROCERY STORE Real Price.sql
@@ -7,6 +7,7 @@ select
 name,
 weight,
 price,
-round(((price*1000.0)/ nullif(weight,0))::numeric,2) as price_per_kg
+round(
+  (price::numeric*1000)/ nullif(weight,0)::numeric,2)::float as price_per_kg
 from products
 order by price_per_kg asc, name asc;
